feat: Refactor some more API routes to actions
