+++ apps/web/package.json
@@ -60,7 +60,6 @@
     "jszip": "^3.10.1",
     "lodash": "^4.17.21",
     "lucide-react": "^0.294.0",
-    "masonic": "^4.1.0",
     "moment": "^2.30.1",
     "next": "14.2.3",
     "next-auth": "^4.24.5",
@@ -76,7 +75,6 @@
     "react-hook-form": "^7.49.2",
     "react-hot-toast": "^2.4.1",
     "react-player": "^2.14.1",
-    "react-responsive-masonry": "^2.7.1",
     "react-rnd": "^10.4.1",
     "react-scroll-parallax": "^3.4.5",
     "react-tooltip": "^5.26.3",
